Move router_gemm_bf16_fp32 to top-level directories
Following PR #3279 review feedback, move files from DSA subdirectories
to top-level directories:
- src/flag_gems/fused/DSA/router_gemm_bf16_fp32.py -> src/flag_gems/fused/router_gemm_bf16_fp32.py
- tests/test_DSA/test_router_gemm_bf16_fp32.py -> tests/test_router_gemm_bf16_fp32.py

Update import in __init__.py accordingly.
